[BugFix] 로그인 후 기능 관련 수정
