+++ src/App.js
@@ -1,4 +1,4 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import Layout from './layouts/Layout';
 import Main from './pages/fairytale/Main'
 import Login from './pages/user/Login';
@@ -9,29 +9,31 @@ import Make from './pages/fairytale/Make'
 import Deelp from './pages/fairytale/Deelp';
 import Clone from './pages/clone/Clone';
 import Info from './pages/mypage/Info'
-
+import { AuthProvider } from './components/auth/AuthContext';
 
 function App() {
 
   return (
-    <BrowserRouter>
+    <AuthProvider>
+      <BrowserRouter>
 
-      <Routes>
+        <Routes>
 
-        <Route path='/' element={<Layout/>}>
-          <Route index element={ <Main/> }/>
-          <Route path='login' element={ <Login/> }/>
-          <Route path='dubing' element={ <Dubing/> }/>
-          <Route path='enroll' element={ <Enroll/> }/>
-          <Route path="make" element={ <Make/> }/>
-          <Route path="deelp" element={ <Deelp/> }/>
-          <Route path="clone" element={ <Clone/> }/>
-          <Route path="info" element={ <Info/> }/>
-        </Route>
+          <Route path='/' element={<Layout/>}>
+            <Route index element={ <Main/> }/>
+            <Route path='login' element={ <Login/> }/>
+            <Route path='dubing' element={ <Dubing/> }/>
+            <Route path='enroll' element={ <Enroll/> }/>
+            <Route path="make" element={ <Make/> }/>
+            <Route path="deelp" element={ <Deelp/> }/>
+            <Route path="clone" element={ <Clone/> }/>
+            <Route path="info" element={ <Info/> }/>
+          </Route>
 
-      </Routes>
+        </Routes>
 
-    </BrowserRouter>
+      </BrowserRouter>
+    </AuthProvider>
   );
 }
